+++ src/main/java/com/peeerr/climbing/dto/request/MemberCreateRequest.java
@@ -8,7 +8,7 @@ import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
 import lombok.*;
 
-import static com.peeerr.climbing.constant.ErrorMessage.*;
+import static com.peeerr.climbing.exception.ErrorMessage.*;
 
 @EqualsAndHashCode
 @NoArgsConstructor
