+++ src/main/java/com/peeerr/climbing/security/handler/Http401Handler.java
@@ -1,7 +1,7 @@
 package com.peeerr.climbing.security.handler;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.peeerr.climbing.constant.ErrorMessage;
+import com.peeerr.climbing.exception.ErrorMessage;
 import com.peeerr.climbing.dto.common.ErrorResponse;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
