+++ src/main/java/com/peeerr/climbing/service/S3FileUploader.java
@@ -8,7 +8,7 @@ import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
-import com.peeerr.climbing.constant.ErrorMessage;
+import com.peeerr.climbing.exception.ErrorMessage;
 import com.peeerr.climbing.dto.FileStoreDto;
 import com.peeerr.climbing.exception.ClimbingException;
 import jakarta.annotation.PostConstruct;
