+++ src/test/java/com/peeerr/climbing/integration/docs/FileDocTest.java
@@ -1,6 +1,6 @@
 package com.peeerr.climbing.integration.docs;
 
-import com.peeerr.climbing.constant.ErrorMessage;
+import com.peeerr.climbing.exception.ErrorMessage;
 import com.peeerr.climbing.security.MemberPrincipal;
 import com.peeerr.climbing.domain.Category;
 import com.peeerr.climbing.repository.CategoryRepository;
@@ -13,7 +13,6 @@ import com.peeerr.climbing.repository.PostRepository;
 import com.peeerr.climbing.domain.Member;
 import com.peeerr.climbing.repository.MemberRepository;
 import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
